requirements.txt
aiogram>=3.4.1
python-dotenv>=1.0.0
aiofiles>=23.2.1
aiohttp>=3.9.0
asyncpg>=0.29.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
redis>=5.0.0  # Для RedisStorage (персистентное хранение FSM)
